Fronend completed of watch wear os

+++ lib/screens/splash_screen.dart
@@ -1,11 +1,14 @@
+import 'package:finance_track/screens/watchwear_notlogin.dart';
+import 'package:finance_track/screens/watchwearos_homescreen.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-
+import 'package:watch_connectivity/watch_connectivity.dart';
 import 'home_screen.dart';
 import 'login_screen.dart';
 
 class SplashScreen extends StatefulWidget {
-  const SplashScreen({Key? key}) : super(key: key);
+  final bool isWear;
+  const SplashScreen({super.key, required this.isWear});
 
   @override
   State<SplashScreen> createState() => _SplashScreenState();
@@ -21,23 +24,55 @@ class _SplashScreenState extends State<SplashScreen> {
   Future<void> _checkLoginStatus() async {
     final prefs = await SharedPreferences.getInstance();
     final isLoggedIn = prefs.getBool('isLoggedIn') ?? false;
+    final isWearOsLoggedIn = prefs.getBool('isWearLoggedIn') ?? false;
     final loginTimestamp = prefs.getInt('loginTimestamp') ?? 0;
 
     final currentTime = DateTime.now().millisecondsSinceEpoch;
     final sixtyMinutes = 60 * 60 * 1000;
 
-    if (isLoggedIn && (currentTime - loginTimestamp) < sixtyMinutes) {
-      Navigator.pushReplacement(
-        context,
-        MaterialPageRoute(builder: (context) => const FinanceHomeScreen()),
-      );
-    } else {
-      await prefs.clear();
-      Navigator.pushReplacement(
-        context,
-        MaterialPageRoute(builder: (context) => const LoginScreen()),
-      );
+    if(widget.isWear)
+      {
+        Navigator.pushReplacement(
+          context,
+          MaterialPageRoute(builder: (context) => const WatchwearosHomescreen()),
+        );
+        // if (isWearOsLoggedIn && (currentTime - loginTimestamp) < sixtyMinutes) {
+        //   Navigator.pushReplacement(
+        //     context,
+        //     MaterialPageRoute(builder: (context) => const WatchwearosHomescreen()),
+        //   );
+        // }
+        // else
+        //   {
+        //     await prefs.clear();
+        //     Navigator.pushReplacement(
+        //       context,
+        //       MaterialPageRoute(builder: (context) => const WatchwearNotlogin()),
+        //     );
+        //   }
+
+
+      }
+    else
+    {
+      if (isLoggedIn && (currentTime - loginTimestamp) < sixtyMinutes)
+      {
+        Navigator.pushReplacement(
+          context,
+          MaterialPageRoute(builder: (context) => const FinanceHomeScreen()),
+        );
+      }
+      else
+        {
+          await prefs.clear();
+          Navigator.pushReplacement(
+            context,
+            MaterialPageRoute(builder: (context) => const LoginScreen()),
+          );
+        }
     }
+
+
   }
 
   @override
